+++ src/index/mod.rs
@@ -21,7 +21,7 @@ use std::path::PathBuf;
 
 pub trait ReadEx {
     fn read_u8(&mut self) -> anyhow::Result<u8>;
-    fn read_u64_be(&mut self) -> anyhow::Result<usize>;
+    fn read_u64_be(&mut self) -> anyhow::Result<u64>;
     fn read_u16_be(&mut self) -> anyhow::Result<u16>;
     fn read_compressed(&mut self) -> anyhow::Result<Vec<u8>>;
 }
@@ -33,10 +33,10 @@ impl<R: Read> ReadEx for R {
         Ok(buf[0])
     }
 
-    fn read_u64_be(&mut self) -> anyhow::Result<usize> {
+    fn read_u64_be(&mut self) -> anyhow::Result<u64> {
         let mut buf = [0u8; 8];
         self.read_exact(&mut buf)?;
-        Ok(u64::from_be_bytes(buf) as usize)
+        Ok(u64::from_be_bytes(buf))
     }
 
     fn read_u16_be(&mut self) -> anyhow::Result<u16> {
@@ -45,7 +45,7 @@ impl<R: Read> ReadEx for R {
         Ok(u16::from_be_bytes(buf))
     }
     fn read_compressed(&mut self) -> anyhow::Result<Vec<u8>> {
-        let mut compressed = vec![0u8; self.read_u64_be()?];
+        let mut compressed = vec![0u8; self.read_u64_be()? as usize];
         self.read_exact(&mut compressed)?;
         Ok(zstd::decode_all(Cursor::new(&compressed))?)
     }
@@ -54,8 +54,8 @@ impl<R: Read> ReadEx for R {
 pub trait WriteEx {
     fn write_u8(&mut self, v: u8) -> anyhow::Result<()>;
     fn write_u16_be(&mut self, v: u16) -> anyhow::Result<()>;
-    fn write_u64_be(&mut self, v: usize) -> anyhow::Result<()>;
-    fn write_compressed(&mut self, data: &[u8]) -> anyhow::Result<usize>;
+    fn write_u64_be(&mut self, v: u64) -> anyhow::Result<()>;
+    fn write_compressed(&mut self, data: &[u8]) -> anyhow::Result<u64>;
 }
 
 impl<W: Write> WriteEx for W {
@@ -69,16 +69,16 @@ impl<W: Write> WriteEx for W {
         Ok(())
     }
 
-    fn write_u64_be(&mut self, v: usize) -> anyhow::Result<()> {
-        self.write(&(v as u64).to_be_bytes())?;
+    fn write_u64_be(&mut self, v: u64) -> anyhow::Result<()> {
+        self.write(&v.to_be_bytes())?;
         Ok(())
     }
 
-    fn write_compressed(&mut self, data: &[u8]) -> anyhow::Result<usize> {
+    fn write_compressed(&mut self, data: &[u8]) -> anyhow::Result<u64> {
         let compressed = zstd::encode_all(data, zstd::DEFAULT_COMPRESSION_LEVEL)?;
-        self.write_u64_be(compressed.len())?;
+        self.write_u64_be(compressed.len() as u64)?;
         self.write(&compressed)?;
-        Ok(8 + compressed.len())
+        Ok(8 + compressed.len() as u64)
     }
 }
